+++ util/kstdlib/src/kstdlib.c
@@ -3,7 +3,11 @@
 #include "kstring.h"
 #include "kernel/include/random.h"
 
-size_t kitoa(sint32 a_value, char *a_buffer, size_t *a_length, size_t a_base)
+uint32 kitoa(
+    sint32 a_value, 
+    char *a_buffer, 
+    size_t *a_length, 
+    size_t a_base)
 {
     if (a_buffer == NULL || a_length == NULL)
     {
@@ -34,7 +38,11 @@ size_t kitoa(sint32 a_value, char *a_buffer, size_t *a_length, size_t a_base)
     return ERROR_SUCCESS;
 }
 
-size_t kutoa(uint32 a_value, char *a_buffer, size_t *a_length, size_t a_base)
+uint32 kutoa(
+    uint32 a_value, 
+    char *a_buffer, 
+    size_t *a_length, 
+    size_t a_base)
 {
     if (a_buffer == NULL || a_length == NULL)
     {
@@ -103,7 +111,11 @@ size_t kutoa(uint32 a_value, char *a_buffer, size_t *a_length, size_t a_base)
     return ERROR_SUCCESS;
 }
 
-size_t ki64toa(sint64 a_value, char *a_buffer, size_t *a_length, size_t a_base)
+uint32 ki64toa(
+    sint64 a_value, 
+    char *a_buffer, 
+    size_t *a_length, 
+    size_t a_base)
 {
     if (a_buffer == NULL || a_length == NULL)
     {
@@ -134,7 +146,11 @@ size_t ki64toa(sint64 a_value, char *a_buffer, size_t *a_length, size_t a_base)
     return ERROR_SUCCESS;
 }
 
-size_t ku64toa(uint64 a_value, char *a_buffer, size_t *a_length, size_t a_base)
+uint32 ku64toa(
+    uint64 a_value, 
+    char *a_buffer, 
+    size_t *a_length, 
+    size_t a_base)
 {
     if (a_buffer == NULL || a_length == NULL)
     {
@@ -205,7 +221,8 @@ size_t ku64toa(uint64 a_value, char *a_buffer, size_t *a_length, size_t a_base)
 
 static size_t g_seed;
 
-size_t krand(size_t *a_value)
+uint32 krand(
+    size_t *a_value)
 {
     if (a_value == NULL)
     {
@@ -221,7 +238,10 @@ size_t krand(size_t *a_value)
     return ERROR_SUCCESS;
 }
 
-size_t kranduint(size_t *a_value, size_t a_start, size_t a_end)
+uint32 kranduint(
+    size_t *a_value, 
+    size_t a_start, 
+    size_t a_end)
 {
     if (a_value == NULL)
     {
@@ -240,7 +260,8 @@ size_t kranduint(size_t *a_value, size_t a_start, size_t a_end)
     return ERROR_SUCCESS;
 }
 
-size_t ksrand(uint64 a_seed)
+uint32 ksrand(
+    uint64 a_seed)
 {
     g_seed = a_seed ^ kernel_random32();
